+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-# [0.6.6] - New Extensions
+# [0.6.7] - New Extensions & Optimizations
+
+#### New methods
+
+- `isSwiftCode` : A static method to check whether the provided `String` is a valid Swift code. [Source](https://en.wikipedia.org/wiki/ISO_9362#:~:text=The%20SWIFT%20code%20is%208,institution%20code%20or%20bank%20code.&text=if%20the%20second%20character%20is,used%20on%20the%20live%20network.)
+- `digitCount` : A static method to count (if any) digits in the `String`
+- `isAscii` : A static method to check whether a `String` is consisted only from ASCII characters
+- `isAnagramOf(String s)` : A static method to check whether the provided `s` `String` is anagram of the extended `String`
+- `isPalindrome` : A static method to check whether the `String` is palindrome
+- `isMixedCase` : A static method to check whether the `String` is consisted of both upper & lower case letters
+
+#### Optimized methods
+
+- Optimizes `replaceGreek` method
+- Optimizes `toGreekUppercase` method
+- Optimizes `getLevenshtein` by using Wagner-Fisher algorithm
+
+## [0.6.6] - New Extensions
 
 - `isIban` : A static method to check whether the provided `String` is a valid IBAN. [Source](https://www.iban.com/structure)
 - `isGreekId` : A static method to check whether the provided `String` is a valid Greek Identification Number. [Source](https://en.wikipedia.org/wiki/National_identification_number#Greece)